# pig-fastapi/app/openapi.yaml
openapi: 3.0.0
info:
  description: API for playing the classic dice game Pig.
  title: Pig Game API
  version: 1.0.0
servers:
- description: Local development server
  url: http://localhost:8000
tags:
- description: Operations related to starting and managing games.
  name: Game Management
- description: Operations related to taking turns and playing the game.
  name: Gameplay
paths:
  /game:
    post:
      operationId: create_new_game
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewGameResponse"
          description: Game created successfully.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Start a new Pig game.
      tags:
      - Game Management
  /game/{game_id}:
    get:
      operationId: get_game_state
      parameters:
      - description: The unique identifier of the game.
        explode: false
        in: path
        name: game_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
          description: Current game state.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Game not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Get the current state of a specific game.
      tags:
      - Game Management
  /game/{game_id}/roll:
    post:
      operationId: roll_die
      parameters:
      - description: The unique identifier of the game.
        explode: false
        in: path
        name: game_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
          description: Game state after rolling the die.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Invalid game state (e.g., game already over)."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Game not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Roll the die for the current player.
      tags:
      - Gameplay
  /game/{game_id}/hold:
    post:
      operationId: hold_turn
      parameters:
      - description: The unique identifier of the game.
        explode: false
        in: path
        name: game_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
          description: Game state after holding.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Invalid game state (e.g., game already over, cannot hold after\
            \ rolling a 1)."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Game not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: "Current player holds, ending their turn and adding turn total to score."
      tags:
      - Gameplay
components:
  schemas:
    GameState:
      description: Represents the current state of a Pig game.
      example:
        scores:
        - 0
        - 0
        turn_total: 12
        is_game_over: false
        winner_player_index: 0
        last_roll: 4
        current_player_index: 0
        game_id: a1b2c3d4-e5f6-7890-1234-567890abcdef
      properties:
        game_id:
          description: Unique identifier for the game.
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
          format: uuid
          readOnly: true
          title: game_id
          type: string
        current_player_index:
          description: Index of the player whose turn it is (0 or 1).
          example: 0
          maximum: 1
          minimum: 0
          title: current_player_index
          type: integer
        scores:
          description: Array of current overall scores for each player.
          example:
          - 0
          - 0
          items:
            minimum: 0
            type: integer
          maxItems: 2
          minItems: 2
          title: scores
          type: array
        turn_total:
          description: Points accumulated in the current players turn.
          example: 12
          minimum: 0
          title: turn_total
          type: integer
        last_roll:
          description: The result of the last die roll. Null if no roll yet this turn.
          example: 4
          maximum: 6
          minimum: 1
          nullable: true
          title: last_roll
          type: integer
        is_game_over:
          description: Indicates if the game has ended.
          example: false
          readOnly: true
          title: is_game_over
          type: boolean
        winner_player_index:
          description: Index of the winning player if the game is over. Null otherwise.
          maximum: 1
          minimum: 0
          nullable: true
          readOnly: true
          title: winner_player_index
          type: integer
          example: null
      required:
      - current_player_index
      - game_id
      - is_game_over
      - scores
      - turn_total
      title: GameState
      type: object
    NewGameResponse:
      description: Response containing the ID of a newly created game.
      example:
        game_id: a1b2c3d4-e5f6-7890-1234-567890abcdef
      properties:
        game_id:
          description: Unique identifier for the new game.
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
          format: uuid
          title: game_id
          type: string
      required:
      - game_id
      title: NewGameResponse
      type: object
    ErrorResponse:
      description: Standard error response format.
      example:
        detail: Game not found
      properties:
        detail:
          description: A detailed error message.
          example: Game not found
          title: detail
          type: string
      required:
      - detail
      title: ErrorResponse
      type: object