# pig-fastapi/app/src/openapi_server/models/game_state.py
# coding: utf-8

"""
    Pig Game API

    API for playing the classic dice game Pig.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictBool
from typing import Any, ClassVar, Dict, List, Optional
from typing_extensions import Annotated
from uuid import UUID
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class GameState(BaseModel):
    """
    Represents the current state of a Pig game.
    """ # noqa: E501
    game_id: UUID = Field(description="Unique identifier for the game.")
    current_player_index: Annotated[int, Field(le=1, strict=True, ge=0)] = Field(description="Index of the player whose turn it is (0 or 1).")
    scores: Annotated[List[Annotated[int, Field(strict=True, ge=0)]], Field(min_length=2, max_length=2)] = Field(description="Array of current overall scores for each player.")
    turn_total: Annotated[int, Field(strict=True, ge=0)] = Field(description="Points accumulated in the current players turn.")
    last_roll: Optional[Annotated[int, Field(le=6, strict=True, ge=1)]] = Field(default=None, description="The result of the last die roll. Null if no roll yet this turn.")
    is_game_over: StrictBool = Field(description="Indicates if the game has ended.")
    winner_player_index: Optional[Annotated[int, Field(le=1, strict=True, ge=0)]] = Field(default=None, description="Index of the winning player if the game is over. Null otherwise.")
    __properties: ClassVar[List[str]] = ["game_id", "current_player_index", "scores", "turn_total", "last_roll", "is_game_over", "winner_player_index"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of GameState from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        * OpenAPI `readOnly` fields are excluded.
        * OpenAPI `readOnly` fields are excluded.
        * OpenAPI `readOnly` fields are excluded.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
                "game_id",
                "is_game_over",
                "winner_player_index",
            },
            exclude_none=True,
        )
        # set to None if last_roll (nullable) is None
        # and model_fields_set contains the field
        if self.last_roll is None and "last_roll" in self.model_fields_set:
            _dict['last_roll'] = None

        # set to None if winner_player_index (nullable) is None
        # and model_fields_set contains the field
        if self.winner_player_index is None and "winner_player_index" in self.model_fields_set:
            _dict['winner_player_index'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of GameState from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "game_id": obj.get("game_id"),
            "current_player_index": obj.get("current_player_index"),
            "scores": obj.get("scores"),
            "turn_total": obj.get("turn_total"),
            "last_roll": obj.get("last_roll"),
            "is_game_over": obj.get("is_game_over"),
            "winner_player_index": obj.get("winner_player_index")
        })
        return _obj